refactor: create a hook useAddressForm
